+++ src/core/config/GOConfigFileReader.h
@@ -10,6 +10,7 @@
 
 #include <wx/string.h>
 
+#include <functional>
 #include <map>
 
 class GOOpenedFile;
@@ -28,6 +29,10 @@ public:
   bool Read(GOOpenedFile *file);
   bool Read(wxString filename);
   wxString GetHash();
+
+  using ProgressFn = std::function<void(unsigned, const wxString &)>;
+  bool ReadWithProgress(
+    GOOpenedFile *file, const wxString &phaseLabel, ProgressFn onProgress);
 
   const std::map<wxString, std::map<wxString, wxString>> &GetContent();
   wxString getEntry(wxString group, wxString name);