+++ src/grandorgue/GOOrganController.h
@@ -105,6 +105,16 @@ private:
 
   GOMemoryPool m_pool;
   GOSoundOrganEngine m_SoundEngine;
+
+  // timing variables for profiling (ms)
+  long long __tim_parse_ms;
+  long long __tim_cmb_ms;
+  long long __tim_readorgan_ms;
+  long long __tim_cache_ms;
+  long long __tim_panels_ms;
+  long long __tim_ranks_ms;
+  long long __tim_modelrest_ms;
+
   GOGuiImageCache *mp_ImageCache;
   GOLabelControl m_PitchLabel;
   GOLabelControl m_TemperamentLabel;
@@ -116,7 +126,7 @@ private:
   // if modified then sets m_IsOrganModified
   void OnIsModifiedChanged(bool modified);
 
-  void ReadOrganFile(GOConfigReader &cfg);
+  void ReadOrganFile(GOConfigReader &cfg, GOProgressMonitor &monitor);
   GOHashType GenerateCacheHash();
   wxString GenerateSettingFileName();
   wxString GenerateCacheFileName();